+++ src/handlers/messageCreate.ts
@@ -4,6 +4,8 @@ import Logger from '../lib/logger';
 import commands from '../commands';
 import { makeEmbed } from '../lib/embed';
 import { client, DEBUG_MODE } from '../index';
+import { CommandDefinition, isExecutorCommand, isMessageCommand } from '../lib/command';
+import { typeCommand } from '../lib/typeCommand';
 
 module.exports = {
     event: 'messageCreate',
@@ -33,31 +35,36 @@ module.exports = {
             const command = commands[usedCommand];
 
             if (command) {
-                const { executor, name, requiredPermissions } = command;
-
+                const { name, requiredPermissions } = command;
                 const commandsArray = Array.isArray(name) ? name : [name];
-
                 const member = await msg.guild.members.fetch(msg.author);
 
                 if (!requiredPermissions || requiredPermissions.every((permission) => member.permissions.has(permission))) {
                     if (commandsArray.includes(usedCommand)) {
-                        try {
-                            await executor(msg, client);
-                            transaction.result = 'success';
-                        } catch ({ name, message, stack }) {
-                            Logger.error({ name, message, stack });
-                            const errorEmbed = makeEmbed({
-                                color: Colors.Red,
-                                title: 'Error while Executing Command',
-                                description: DEBUG_MODE ? `\`\`\`D\n${stack}\`\`\`` : `\`\`\`\n${name}: ${message}\n\`\`\``,
-                            });
-
-                            await msg.channel.send({ embeds: [errorEmbed] });
-
-                            transaction.result = 'error';
+                        let executor;
+                        if (isExecutorCommand(command)) {
+                            ({ executor } = (command as CommandDefinition));
+                        } else if (isMessageCommand(command)) {
+                            executor = typeCommand.executor;
                         }
+                        if (executor) {
+                            try {
+                                await executor(msg, client);
+                                transaction.result = 'success';
+                            } catch ({ name, message, stack }) {
+                                Logger.error({ name, message, stack });
+                                const errorEmbed = makeEmbed({
+                                    color: Colors.Red,
+                                    title: 'Error while Executing Command',
+                                    description: DEBUG_MODE ? `\`\`\`D\n${stack}\`\`\`` : `\`\`\`\n${name}: ${message}\n\`\`\``,
+                                });
 
-                        Logger.debug('Command executor done.');
+                                await msg.channel.send({ embeds: [errorEmbed] });
+
+                                transaction.result = 'error';
+                            }
+                            Logger.debug('Command executor done.');
+                        }
                     }
                 } else {
                     await msg.reply(`you do not have sufficient permissions to use this command. (missing: ${requiredPermissions.join(', ')})`);